separate epi readout + tests
