+++ src/mrseq/utils/__init__.py
@@ -3,3 +3,4 @@ from mrseq.utils.system_defaults import sys_defaults
 from mrseq.utils.vds import variable_density_spiral_trajectory
 from mrseq.utils.trajectory import spiral_acquisition
 from mrseq.utils.ismrmrd import create_header, combine_ismrmrd_files
+from mrseq.utils.EpiReadout import EpiReadout
